refactor(web): colocate dashboard layout under route as -layout/
Move apps/web/src/layout/dashboard/ into the routes tree so the shell
lives next to its only consumer (_dashboard.tsx). Dashboard shell goes
to routes/_authenticated/$slug/$teamSlug/-layout/ (with hooks/ and
-sidebar-scope-switcher/ inside); settings shell goes to
_dashboard/settings/-layout/. The - prefix keeps TanStack Router from
treating these as routes. apps/web/src/layout/ is deleted.
