+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/early-access-sidebar-banner.tsx
