+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/hooks/use-sidebar-tabs.ts
