+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/keyboard-shortcuts-sheet.tsx
