+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/sidebar-account-menu.tsx
