+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/sidebar-browse-tabs.tsx
