+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/sidebar-chat-panel.tsx
