+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/sidebar-scope-switcher.tsx
