+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/theme-switcher.tsx
