+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/settings.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, Outlet } from "@tanstack/react-router";
-import { SettingsLayout } from "@/layout/dashboard/ui/settings-layout";
+import { SettingsLayout } from "./settings/-layout/settings-layout";
 
 export const Route = createFileRoute(
    "/_authenticated/$slug/$teamSlug/_dashboard/settings",
