+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/settings/-layout/settings-addon-gated-page.tsx
