+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/settings/index.tsx
@@ -1,6 +1,6 @@
 import { useMediaQuery } from "foxact/use-media-query";
 import { createFileRoute, Navigate } from "@tanstack/react-router";
-import { SettingsMobileNav } from "@/layout/dashboard/ui/settings-mobile-nav";
+import { SettingsMobileNav } from "./-layout/settings-mobile-nav";
 
 export const Route = createFileRoute(
    "/_authenticated/$slug/$teamSlug/_dashboard/settings/",
